Mysecrets enhancements

+++ bashrc.mysecrets.bash
@@ -8,8 +8,10 @@ mysecretsfile() {
 
 mysecrets() {
     if [ -f ~/.mysecrets ] && [ -f "$(mysecretsfile)" ]; then
-        gpg -d $(mysecretsfile) 2> /dev/null || return $?
-        echo
+        local pass
+        read -sp $'Password:\n' pass
+        { echo "${pass}" | gpg -d --batch --yes --passphrase-fd 0 "$(mysecretsfile)" 2> /dev/null; } || return $?
+        echo;
     else
         echo "Invalid config (~/.mysecrets)"
         return 1
@@ -17,20 +19,32 @@ mysecrets() {
     return 0
 }
 
-editsecrest() {
-    if [ -f ~/.mysecrets ]; then
-        TMPFILE=$(TMPDIR=/dev/shm mktemp) || return $?
-        TMPOUT="$TMPFILE.gpg"
-        trap "rm -rf $TMPOUT $TMPFILE" EXIT
-        mysecrets >"$TMPFILE" || return $?
-        vi $TMPFILE || return $?
-        gpg --output $TMPOUT --symmetric $TMPFILE || return $?
+mysecrets-edit() {
+    if [ -f ~/.mysecrets ] && [ -f "$(mysecretsfile)" ]; then
+        # Prepare working dir
+        local tmpfile="$(TMPDIR=/dev/shm mktemp)" || return $?
+        local tmpout="${tmpfile}.gpg"
+
+        # Decrypt
+        local pass
+        read -sp $'Password:\n' pass
+        { echo "${pass}" | gpg -d --batch --yes --passphrase-fd 0 "$(mysecretsfile)" > "${tmpfile}" 2> /dev/null; } || return $?
+
+        # Edit
+        edit ${tmpfile} || return $?
+
+        # Re-encrypt
+        { echo "${pass}" | gpg --batch --yes --passphrase-fd 0 --output ${tmpout} --symmetric ${tmpfile}; } || return $?
         cp $(mysecretsfile) "$(mysecretsfile).bak" || return $?
-        mv $TMPOUT $(mysecretsfile) || return $?
-        return 0
+        mv ${tmpout} $(mysecretsfile) || return $?
     else
-        echo "Secrets file not found"
+        echo "Invalid config (~/.mysecrets)"
         return 1
     fi
-
+    return 0
 }
+
+mysecrets-token() {
+	mysecrets | grep -oP --color=auto "(?<=${1} ).*"
+}
+
